backend/requirements.txt
# Core FastAPI and dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
websockets==15.0.1
starlette==0.47.2

# HTTP Tools
httptools==0.6.4
httpcore==1.0.9
httpx==0.28.1
h11==0.16.0
requests==2.32.5
urllib3==2.5.0

# Database
sqlalchemy==2.0.43
alembic==1.16.4
greenlet==3.2.4

# Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Rate limiting
slowapi==0.1.9

# Torrent handling
libtorrent==2.0.11

# Utilities
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
aiofiles==24.1.0
psutil==5.9.8

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
mypy_extensions==1.1.0

# Production
gunicorn==21.2.0

# Core dependencies
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
annotated-types==0.7.0
